requirements.txt
langchain
huggingface-hub
langchain-community
langchain-groq
functions-framework
python-dotenv
openai